requirements.txt
api==0.0.7
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
easydict==1.9
filelock==3.3.2
Flask==2.0.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
fonttools==4.25.0
greenlet==1.1.2
gunicorn==20.1.0
huggingface-hub==0.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
munkres==1.1.4
networkx==2.6.3
nltk==3.6.6
nose==1.3.7
numpy==1.21.4
pandas==1.3.5
Pillow==8.4.0
pytz==2021.3
PyWavelets==1.2.0
PyYAML==5.4.1
regex==2021.11.2
requests==2.26.0
sacremoses==0.0.46
scikit-image==0.19.1
scipy==1.7.3
sentencepiece==0.1.96
SQLAlchemy==1.4.29
sqlparse==0.4.2
tifffile==2021.11.2
tokenizers==0.10.3
tqdm==4.62.3
transformers==4.12.3
urllib3==1.26.7
Werkzeug==2.0.2
wincertstore==0.2
